refactor: isolate rope diagnostics
